requirements.txt
django-admin-thumbnails==0.2.2
